+++ projects/rocprofiler-compute/src/argparser.py
@@ -1030,20 +1030,6 @@ Examples:
     # Experimental Features
     ## ----------------------------
     analyze_group.add_argument(
-        "--membw-analysis",
-        dest="membw_analysis",
-        required=False,
-        default=False,
-        base_action="store_const",
-        action=ExperimentalAction,
-        experimental_enabled=experimental_enabled,
-        feature_label="Memory Bandwidth Analysis",
-        nargs=0,
-        const=True,
-        help="\t\tEnable block 30 (memory bandwidth specific) for analysis mode.",
-    )
-
-    analyze_group.add_argument(
         "--gui",
         type=int,
         nargs="?",